Add fit and predict api endpoints

+++ bee.py
@@ -15,36 +15,40 @@ basedir = os.path.abspath(os.path.dirname(__file__))
 ALGORITHMS = {'random-forest': RandomForest, 'svm': SVM}
 
 
-def extract_and_read_source_files(train_filename, test_filename=None):
-    print 'Extracting {}'.format(train_filename)
+def extract_and_read_source_files(train_filename=None, test_filename=None):
+    ret = None
 
-    os.system('mkdir train_files')
-    os.system('unzip -q {} -d train_files'.format(train_filename))
+    if train_filename:
+        train_filename = 'files/{}'.format(train_filename)
 
-    train_source_code = []
-    train_labels = []
-    ast_nodes = []
+        os.system('mkdir train_files')
+        os.system('unzip -q {} -d train_files'.format(train_filename))
 
-    for filename in sorted(os.listdir('train_files')):
-        filename_split = filename.split('_')
-        if 'cpp' not in filename_split[-1] and 'cxx' not in filename_split[-1]:
-            continue
+        train_source_code = []
+        train_labels = []
+        ast_nodes = []
+
+        for filename in sorted(os.listdir('train_files')):
+            filename_split = filename.split('_')
+            if 'cpp' not in filename_split[-1] and 'cxx' not in filename_split[-1]:
+                continue
+            
+            author = "{} {}".format(filename_split[0].encode('utf-8'), filename_split[1].encode('utf-8'))
+            train_labels.append(author)
+            with open(basedir + '/train_files/' + filename, 'r') as f:
+                train_source_code.append(f.read())
+
+            #ast_nodes.append(subprocess.check_output(['java', '-jar', 'CodeSensor.jar', basedir + '/dataset/' + filename]))
+
         
-        author = "{} {}".format(filename_split[0].encode('utf-8'), filename_split[1].encode('utf-8'))
-        train_labels.append(author)
-        with open(basedir + '/train_files/' + filename, 'r') as f:
-            train_source_code.append(f.read())
+        ret = train_source_code, train_labels
+        os.system('rm -rf train_files')
 
-        #ast_nodes.append(subprocess.check_output(['java', '-jar', 'CodeSensor.jar', basedir + '/dataset/' + filename]))
-
-    
-    os.system('rm -rf train_files')
-
-    print 'Extracting {}'.format(test_filename)
-    
     test_source_code = []
     test_labels = []
     if test_filename:
+        test_filename = 'files/{}'.format(test_filename)
+
         os.system('mkdir test_files')
         os.system('unzip -q {} -d test_files'.format(test_filename))
 
@@ -60,10 +64,12 @@ def extract_and_read_source_files(train_filename, test_filename=None):
         
         os.system('rm -rf test_files')
 
-    return train_source_code, train_labels, test_source_code, test_labels
+        ret = ret + (test_source_code, test_labels)
+    
+    return ret
 
 
-def train(ml_algorithm, train_files, train_labels, test_files, test_labels):
+def train_and_predict(ml_algorithm, train_files, train_labels, test_files, test_labels):
     train_features, test_features = FeatureExtractor.get_features(train_files, test_files)
 
     try:
@@ -117,7 +123,8 @@ def train_kfold(ml_algorithm, source_code, labels, ast_nodes=None):
             train_features, test_features = FeatureExtractor.get_features(source_code[train_indices],
                                                                             source_code[test_indices])
             
-            score = mla.train(train_features, test_features, labels[train_indices], labels[test_indices])
+
+            score = mla.fit_and_predict(train_features, test_features, labels[train_indices], labels[test_indices])
             print 'Score after {}th fold is {}'.format(i, score)
             accuracy += score
                 
@@ -126,6 +133,34 @@ def train_kfold(ml_algorithm, source_code, labels, ast_nodes=None):
         print 'Execution time: {}'.format(end)
         print '-----------------------------------------------'
 
+
+def fit(ml_algorithm, train_files, train_labels, name):
+    train_features, _ = FeatureExtractor.get_features(train_files, [])
+
+    try:
+        algorithm_type = ALGORITHMS[ml_algorithm]
+    except KeyError:
+        print 'Algorithm type not valid!'
+        return
+    
+    mla = algorithm_type()
+
+    mla.fit(train_features, train_labels, name)
+
+
+def get_probs(ml_algorithm, train_files, test_files, name):
+    _, test_features = FeatureExtractor.get_features(train_files, test_files)
+
+    try:
+        algorithm_type = ALGORITHMS[ml_algorithm]
+    except KeyError:
+        print 'Algorithm type not valid!'
+        return
+    
+    mla = algorithm_type()
+
+    prob_ind, highest_score = mla.predict(test_features, name)
+    return prob_ind, highest_score
 
 
 if __name__ == '__main__':
@@ -136,13 +171,10 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
 
-    train_filename = 'files/{}'.format(args.train_set)
-    test_filename = 'files/{}'.format(args.test_set)
-
-    if test_filename:
-        train_source_code, train_labels, test_source_code, test_labels = extract_and_read_source_files(train_filename, test_filename)
-        train('random-forest', train_source_code, train_labels, test_source_code, test_labels)
+    if args.test_set:
+        train_source_code, train_labels, test_source_code, test_labels = extract_and_read_source_files(args.train_set, args.test_set)
+        train_and_predict('random-forest', train_source_code, train_labels, test_source_code, test_labels)
     
     else:
-        train_source_code, train_labels, _, _ = extract_and_read_source_files(train_filename)
+        train_source_code, train_labels = extract_and_read_source_files(args.train_set)
         train_kfold('random-forest', train_source_code, train_labels)
